chore: only wrap reusable streams
If the first stream is writable and the last stream is readable, that
means the pipeline acts as one single transform which is why we wrap it
in one. Previously we only wrapped if the last stream was readable. But
this ignored the unnessesary wrappigng of when the first stream is not
writable.